extract SampleUploadDisplay.java
