+++ datamanager-app/src/main/java/life/qbic/datamanager/views/projects/project/samples/registration/batch/RegisterSampleBatchDialog.java
@@ -1,7 +1,6 @@
 package life.qbic.datamanager.views.projects.project.samples.registration.batch;
 
 import com.vaadin.flow.component.ClickEvent;
-import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.button.Button;
@@ -22,8 +21,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 import life.qbic.application.commons.ApplicationException;
 import life.qbic.application.commons.FileNameFormatter;
@@ -39,6 +36,9 @@ import life.qbic.datamanager.views.general.download.DownloadComponent;
 import life.qbic.datamanager.views.general.upload.ContentUploadComponent;
 import life.qbic.datamanager.views.general.upload.UploadedFilesChangeListener.FileEntry;
 import life.qbic.datamanager.views.notifications.MessageSourceNotificationFactory;
+import life.qbic.datamanager.views.projects.project.samples.registration.batch.SampleUploadDisplay.InProgressDisplay;
+import life.qbic.datamanager.views.projects.project.samples.registration.batch.SampleUploadDisplay.InvalidUploadDisplay;
+import life.qbic.datamanager.views.projects.project.samples.registration.batch.SampleUploadDisplay.ValidUploadDisplay;
 import life.qbic.logging.api.Logger;
 import life.qbic.logging.service.LoggerFactory;
 import life.qbic.projectmanagement.application.ValidationResult;
@@ -110,7 +110,7 @@ public class RegisterSampleBatchDialog extends WizardDialogWindow {
     contentUploadComponent.setMaxFileSize(DataSize.ofBytes(MAX_FILE_SIZE));
 
     var uploadDisplay = new SampleUploadDisplay();
-    Registration controllerRegistration = new SampleUploadDisplayController(projectId,
+    Registration controllerRegistration = new SampleRegistrationUploadDisplayController(projectId,
         experimentId)
         .control(uploadDisplay, contentUploadComponent);
     uploadDisplay.addDetachListener(it -> controllerRegistration.remove());
@@ -359,154 +359,6 @@ public class RegisterSampleBatchDialog extends WizardDialogWindow {
     fireEvent(new CancelEvent(this, clickEvent.isFromClient()));
   }
 
-  static class InvalidUploadDisplay extends Div {
-
-    public InvalidUploadDisplay(String error) {
-      addClassName("uploaded-item");
-      var fileIcon = VaadinIcon.FILE.create();
-      fileIcon.addClassName("file-icon");
-      Div validationBox = new Div();
-      validationBox.addClassName("validation-display-box");
-      var box = new Div();
-      var failuresTitle = new Span(error);
-      var errorIcon = VaadinIcon.CLOSE_CIRCLE.create();
-      errorIcon.addClassName("error");
-      var header = new Span(errorIcon, failuresTitle);
-      header.addClassName("header");
-      box.add(header);
-      validationBox.add(box);
-      add(validationBox);
-    }
-
-    public InvalidUploadDisplay(String fileName, List<String> failureReasons) {
-      addClassName("uploaded-item");
-      var fileIcon = VaadinIcon.FILE.create();
-      fileIcon.addClassName("file-icon");
-      Span fileNameLabel = new Span(fileIcon, new Span(fileName));
-      fileNameLabel.addClassName("file-name");
-      Div validationBox = new Div();
-      validationBox.addClassName("validation-display-box");
-      var box = new Div();
-      var failuresTitle = new Span("Invalid sample metadata");
-      var errorIcon = VaadinIcon.CLOSE_CIRCLE.create();
-      errorIcon.addClassName("error");
-      var header = new Span(errorIcon, failuresTitle);
-      header.addClassName("header");
-      var instruction = new Span(
-          "Please correct the entries in the uploaded file and re-upload the file.");
-      instruction.addClassName("secondary");
-      Div validationDetails = new Div();
-
-      Map<String, Integer> frequencyMap = failureReasons.stream()
-          .distinct()
-          .collect(Collectors.toMap(
-              Function.identity(),
-              v -> Collections.frequency(failureReasons, v)
-          ));
-      frequencyMap.forEach(
-          (key, frequency) -> {
-            String s = frequency + " sample" + ((frequency > 1) ? "s." : ".");
-            Span span = new Span(s);
-            span.addClassName("bold");
-            validationDetails.add(new Div(new Span(key + " for "), span));
-          });
-      box.add(header, validationDetails, instruction);
-      validationBox.add(box);
-      add(fileNameLabel, validationBox);
-    }
-
-    public InvalidUploadDisplay(String fileName, String failureReason) {
-      addClassName("uploaded-item");
-      var fileIcon = VaadinIcon.FILE.create();
-      fileIcon.addClassName("file-icon");
-      Span fileNameLabel = new Span(fileIcon, new Span(fileName));
-      fileNameLabel.addClassName("file-name");
-      Div validationBox = new Div();
-      validationBox.addClassName("validation-display-box");
-      var box = new Div();
-      var failuresTitle = new Span("Invalid sample metadata");
-      var errorIcon = VaadinIcon.CLOSE_CIRCLE.create();
-      errorIcon.addClassName("error");
-      var header = new Span(errorIcon, failuresTitle);
-      header.addClassName("header");
-      var instruction = new Span(
-          "Please correct the entries in the uploaded file and re-upload the file.");
-      instruction.addClassName("secondary");
-      Div validationDetails = new Div();
-
-      validationDetails.add(new Div(failureReason));
-      box.add(header, validationDetails, instruction);
-      validationBox.add(box);
-      add(fileNameLabel, validationBox);
-    }
-  }
-
-  static class ValidUploadDisplay extends Div {
-
-    private ValidUploadDisplay(String fileName, int count) {
-      addClassName("uploaded-item");
-      var fileIcon = VaadinIcon.FILE.create();
-      fileIcon.addClassName("file-icon");
-      Span fileNameLabel = new Span(fileIcon, new Span(fileName));
-      fileNameLabel.addClassName("file-name");
-      Div validationBox = new Div();
-      validationBox.addClassName("validation-display-box");
-      var box = new Div();
-      var approvedTitle = new Span("Your data has been approved");
-      var validIcon = VaadinIcon.CHECK_CIRCLE_O.create();
-      validIcon.addClassName("success");
-      var header = new Span(validIcon, approvedTitle);
-      header.addClassName("header");
-      var instruction = new Span("Please click Register to register your samples");
-      instruction.addClassName("secondary");
-      Div validationDetails = new Div();
-      var approvedSamples = new Span("%d samples".formatted(count));
-      approvedSamples.addClassName("bold");
-      validationDetails.add(new Span("Sample data for "), approvedSamples,
-          new Span(" is now ready to be registered."));
-      box.add(header, validationDetails, instruction);
-      validationBox.add(box);
-      add(fileNameLabel, validationBox);
-    }
-
-    private ValidUploadDisplay(String fileName) {
-      addClassName("uploaded-item");
-      var fileIcon = VaadinIcon.FILE.create();
-      fileIcon.addClassName("file-icon");
-      Span fileNameLabel = new Span(fileIcon, new Span(fileName));
-      fileNameLabel.addClassName("file-name");
-      Div validationBox = new Div();
-      validationBox.addClassName("validation-display-box");
-      var box = new Div();
-      var approvedTitle = new Span("Your data has been approved");
-      var validIcon = VaadinIcon.CHECK_CIRCLE_O.create();
-      validIcon.addClassName("success");
-      var header = new Span(validIcon, approvedTitle);
-      header.addClassName("header");
-      var instruction = new Span("Please click Register to register your samples");
-      instruction.addClassName("secondary");
-      Div validationDetails = new Div();
-      validationDetails.add(new Span("Sample data is now ready to be registered."));
-      box.add(header, validationDetails, instruction);
-      validationBox.add(box);
-      add(fileNameLabel, validationBox);
-    }
-  }
-
-  private static class InProgressDisplay extends Div {
-
-    private InProgressDisplay(String fileName) {
-      addClassName("uploaded-item");
-      var fileIcon = VaadinIcon.FILE.create();
-      fileIcon.addClassName("file-icon");
-      Span fileNameLabel = new Span(fileIcon, new Span(fileName));
-      fileNameLabel.addClassName("file-name");
-      ProgressBar progressBar = new ProgressBar();
-      progressBar.setIndeterminate(true);
-      add(fileNameLabel, new Div("Validating file..."), progressBar);
-    }
-  }
-
   public static class ConfirmEvent extends ComponentEvent<RegisterSampleBatchDialog> {
 
     private final String batchName;
@@ -555,94 +407,12 @@ public class RegisterSampleBatchDialog extends WizardDialogWindow {
     }
   }
 
-  static class SampleUploadDisplay extends Div {
-
-    private final Map<String, Component> displayedFiles = new HashMap<>();
-    private final InvalidUploadDisplay emptyDisplay = new InvalidUploadDisplay(
-        "Nothing was uploaded. Please upload the sample metadata and try again.");
-
-    private final Div errorArea = new Div();
-    private final Div displayContainer = new Div();
-    private final Div fileSizeRestriction;
-    private final Span displayContainerTitle = new Span("Uploaded File");
-    private DataSize maxFileSize = null;
-
-    public SampleUploadDisplay() {
-      addClassName("display-contents");
-      errorArea.addClassNames("error-message-box");
-      displayContainer.addClassNames("uploaded-items-section");
-
-      displayContainerTitle.addClassNames("section-title");
-
-      var restrictions = new Div();
-      restrictions.addClassNames("restrictions");
-      fileSizeRestriction = new Div();
-      setMaxFileSize(maxFileSize);
-      restrictions.add(fileSizeRestriction);
-
-      updateVisibility();
-      displayContainer.add(emptyDisplay);
-      add(displayContainerTitle, errorArea, displayContainer, restrictions);
-    }
-
-    void updateVisibility() {
-      displayContainer.setVisible(hasContent());
-      displayContainerTitle.setVisible(hasContent());
-
-    }
-
-    void setMaxFileSize(DataSize maxFileSize) {
-      Optional.ofNullable(maxFileSize)
-          .ifPresent(fileSize -> {
-            this.maxFileSize = fileSize;
-            fileSizeRestriction.setText("Maximum file size: " + fileSize);
-          });
-    }
-
-    private boolean hasContent() {
-      return !displayedFiles.isEmpty();
-    }
-
-    void removeDisplay(List<String> fileNames) {
-      List<Component> associatedComponents = fileNames.stream()
-          .map(name -> displayedFiles.getOrDefault(name, null))
-          .filter(Objects::nonNull)
-          .toList();
-      displayContainer.remove(associatedComponents);
-      fileNames.forEach(displayedFiles::remove);
-      if (hasContent()) {
-
-        displayContainer.remove(emptyDisplay);
-      } else {
-        displayContainer.add(emptyDisplay);
-      }
-      updateVisibility();
-    }
-
-    void setDisplay(String fileName, Component display) {
-      Optional<Component> existingComponent = Optional.ofNullable(
-          displayedFiles.getOrDefault(fileName, null));
-      existingComponent.ifPresentOrElse(
-          existing -> displayContainer.replace(existing, display),
-          () -> displayContainer.add(display));
-      displayedFiles.put(fileName, display);
-      displayContainer.remove(emptyDisplay);
-      updateVisibility();
-    }
-
-    public void setErrorText(String errorMessage) {
-      this.errorArea.setText(errorMessage);
-      this.errorArea.setVisible(errorMessage != null && !errorMessage.isBlank());
-
-    }
-  }
-
-  class SampleUploadDisplayController {
+  class SampleRegistrationUploadDisplayController {
 
     private final String projectId;
     private final String experimentId;
 
-    private SampleUploadDisplayController(String projectId, String experimentId) {
+    private SampleRegistrationUploadDisplayController(String projectId, String experimentId) {
       this.projectId = Objects.requireNonNull(projectId);
       this.experimentId = Objects.requireNonNull(experimentId);
     }
